New action "move-contacts" supports moving contacts within the hierarchy (#195)
It is not straight-forward to move contacts from one position in the hierarchy to another because many copies of this hierarchy exist in the data. This new move-contacts action can be used to assign a new parent to contacts. The action downloads the documents that need to be updated, updates the required lineages within those documents, and saves the updated documents to disk. To commit the changes, run the upload-docs action.

#172 

Roughly these are the logical steps used to "move" a contact:

Confirm the move does't violate any rule set by configurable hierarchy (>v3.7)
Confirm the primary contact for any place in the hierarchy isn't being removed from under that place.
Download and update the lineage for the contacts being moved and all descendants of each contact.
Download the reports created by any of the descendant contacts, update the lineage in these reports.
Write all changes to disk.
Run the upload-docs command to commit changes to the server.
The immediate need for this script is being driven by needs for the MoH Nepal project. I've added @mukesh2006 to this review to help confrim that this script meets all of the needs of the MoH project.

The logic in this script is based on similar scripts written by @nomulex and @rosteve. The hope is that this tool can serve to replace those scripts, so adding you guys here to confirm that this meets the needs of your projects.

I've tested this on v2.13, v2.16, v3.4, and master of the webapp. With the permission of the LG team, I've tested this at scale by moving districts to different positions in the hierarchy (~9000 contacts) on the kisii-an.dev instance (v2.16).

The issue #172 mentions providing an overview of the users which will be impacted when documents are moved. Since it is a non-trivial feature add, I've opened #197 to address this in a separate PR.