New --archive flag saves a configuration to disk in ready-to-upload format (#232)
The requirements for medic-conf which arise from medic/medic#5914 are that we need to

Run medic-conf during the build process. This means running medic-conf when API is not yet available.
Any config documents that would be uploaded to API, medic-conf should instead save them to disk (including attachments).
This proposed change introduces a new "repository" object. A repository is injected into each action when they execute. Actions should interact with the target instance through this interface instead of directly creating database connections, or directly issuing requests to an API instance. This is the standard practice of a data-access layer. The ServerRepository implementation provides parity for all existing functionality used in medic-conf via --local, --url, or --instance.

The user can now provide --archive. When the archive option is specified, the data-access layer makes some assumptions about the state of the API and writes uploaded documents to disk. Not all actions are supported with the --archive flag, and users may see an exception "not supported in --archive mode" if they run an unsupported action (example move-contacts, create-user).

The logic which creates the database connection has been moved into a "repository factory" module. This instantiates the repository based on the connection data provided in the command-line arguments.

A new command-line option --source=<path to project> has been added. It defaults to the working directory to avoid any breaking change, but can be set for convenience and improved scriptability.