Tasks and Targets v2 - Declarative configuration update to use new expected schema (#274)
This is an associated change from medic/cht-core#5657.

The ongoing Task and Target Improvements (Rules-Engine v2) which are happening in the 3.8 release require additional data about each task emitted by partner code. After upgrading to 3.8, the task and analytics tabs won't function until partner code is updated to send this new data.

The data can be included without affecting the behavior of the existing webapp. Therefore, we can make a change now to include the data in declarative partner configurations. The goal is that it won't affect anything in ongoing operations and projects will be able to seemlessly upgrade to 3.8 if they have deployed their configuration with medic-conf >3.1.0